pyproject.toml
[project]
name = "lumina"
version = "1.0.0"
description = "Lumina AI Workflow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "fastapi-cli>=0.0.4",
    "gradio>=5.49.1",
    "ipykernel>=6.30.1",
    "ipython>=9.6.0",
    "matplotlib>=3.10.6",
    "openai>=2.2.0",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.0",
    "uvicorn[standard]>=0.37.0",
]

[dependency-groups]
dev = [
    # Development tools (not critical for running the application)
]


